more formatting to react paradigm
