+++ src/containers/Search/Search.js
@@ -1,10 +1,46 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { AiOutlineSearch } from 'react-icons/ai';
 import SearchPokemon from '../../components/search/SearchPokemon'
 import './Search.css';
 
 export default function Search() {
     let [typedSearch, setTypedSearch] = useState("");
+    const [data, setData] = useState([]);
+  
+    const getDescription = (json) => {
+      fetch(`https://pokeapi.co/api/v2/pokemon-species/${json.name}`)
+        .then(res => res.json())
+        .then(res => {console.log("aqui ",res);
+  
+          setData({
+            id: json.id,
+            name: json.name,
+            image: json.sprites.front_default,
+            arrayText: res.flavor_text_entries
+          });
+        })
+        .catch(err => {
+          console.log("ability", err);
+        })
+    }
+  
+    const getPokemon = (search) => {
+      fetch(`https://pokeapi.co/api/v2/pokemon/${search}`)
+        .then(res => res.json())
+        .then(json => {console.log(json)
+          setData({
+            id: json.id,
+            name: json.name,
+            image: json.sprites.front_default
+          });
+          getDescription(json)
+        })
+        .catch(err => console.log(err));
+    }
+
+    useEffect(() => {
+        getPokemon(typedSearch)
+      }, [typedSearch]);
 
     return (
         <><div className="searchBar">
@@ -24,6 +60,6 @@ export default function Search() {
                 </tr>
             </table>
         </div>
-      { typedSearch && <div className="searchResults"><SearchPokemon search={typedSearch} /></div> }</>
+      { typedSearch && <div className="searchResults"><SearchPokemon data={data} /></div> }</>
     )
 }
